Everything seams to be working correctly, repo and clone files get deleted after test is running.  Should have 100% test coverage.
