+++ test/util/cleanupTestApp.js
@@ -1,15 +1,10 @@
-const fs = require('fs-extra')
+const fs = require('fs')
 
 module.exports = function (appDir, next) {
   // use regexp to check appDir included 'test/app', ensure nothing deleted outside Roosevelt test folder
   if (/test[\\/]app/.test(appDir)) {
-    fs.remove(appDir, (err) => {
-      if (err) {
-        next(err)
-      } else {
-        next()
-      }
-    })
+    fs.rmSync(appDir, { recursive: true, force: true })
+    next()
   } else {
     next(new Error(`Directory ${appDir} is not a test app and will not be deleted.`))
   }
