+++ test/util/fallbackSB06.js
@@ -1,0 +1,180 @@
+module.exports = fallbackDependancySandBox
+
+function fallbackDependancySandBox (appDir) {
+  const fs = require('fs')
+  const path = require('path')
+  const testSrc = path.resolve(__dirname, '../../test')
+  const { execSync } = require('child_process')
+  const repoList = ['repo16', 'repo17', 'repo18']
+
+  if (!appDir) {
+    let processEnv
+    if (fs.existsSync(path.join(process.cwd(), 'node_modules')) === false) {
+      processEnv = process.cwd()
+    } else {
+      processEnv = undefined
+    }
+    appDir = processEnv
+  }
+
+  // Checks if the repos folder exist, if not it creates it
+  try {
+    if (!fs.existsSync(`${testSrc}/repos`)) {
+      // creat the repos folder in the test folder
+      fs.mkdirSync(`${testSrc}/repos`, err => {
+        if (err) {
+          console.error(err)
+        }
+        // file written successfully
+      })
+    }
+    // Checks if the clone folder exist, if not it creates it
+    if (!fs.existsSync(`${testSrc}/clones`)) {
+      // creat the clones folder in the test folder
+      fs.mkdirSync(`${testSrc}/clones`, err => {
+        if (err) {
+          console.error(err)
+        }
+        // file written successfully
+      })
+    }
+
+    const repo16Package = {
+      devDependencies: {
+        'fallback-dependencies': '../../../'
+      },
+      fallbackDependencies: {
+        dir: 'lib',
+        reposFile: 'reposFile.json'
+      },
+      scripts: {
+        postinstall: 'node node_modules/fallback-dependencies/fallback-dependencies.js'
+      }
+    }
+    const repo16PackageLocked = {
+      name: 'repo16',
+      lockfileVersion: 3,
+      requires: true,
+      packages: {
+        '': {
+          hasInstallScript: true,
+          devDependencies: {
+            'fallback-dependencies': '../../../'
+          }
+        },
+        '../../..': {
+          version: '0.1.0',
+          dev: true,
+          license: 'CC-BY-16.0',
+          devDependencies: {}
+        },
+        'node_modules/fallback-dependencies': {
+          resolved: '../../..',
+          link: true
+        }
+      }
+    }
+    const repo16FileData = {
+      'fallback-deps-test-repo-17': [
+        '../../../repos/repo17', 'git://github.com/rooseveltframework/roosevelt.git -b 0.21.0 -skip-deps'
+      ]
+    }
+    const repo17Package = {
+      devDependencies: {
+        'fallback-dependencies': '../../../../../../'
+      },
+      fallbackDependencies: {
+        dir: 'lib',
+        repos: {
+          'fallback-deps-test-repo-18': '../../../../../repos/repo18'
+        },
+        reposFile: 'reposFile.json'
+      },
+      scripts: {
+        postinstall: 'node node_modules/fallback-dependencies/fallback-dependencies.js'
+      }
+    }
+    const repo17PackageLocked = {
+      name: 'repo16',
+      lockfileVersion: 3,
+      requires: true,
+      packages: {
+        '': {
+          hasInstallScript: true,
+          devDependencies: {
+            'fallback-dependencies': '../../../../../'
+          }
+        },
+        '../../../../..': {
+          version: '0.1.0',
+          dev: true,
+          license: 'CC-BY-16.0',
+          devDependencies: {}
+        },
+        'node_modules/fallback-dependencies': {
+          resolved: '../../../../..',
+          link: true
+        }
+      }
+    }
+    const repo18Package = {}
+    const repo18PackageLocked = {
+      name: 'repo18',
+      lockfileVersion: 3,
+      requires: true,
+      packages: {}
+    }
+
+    const packageList = [[repo16Package, repo16PackageLocked], [repo17Package, repo17PackageLocked], [repo18Package, repo18PackageLocked]]
+    for (const id in repoList) {
+      if (!fs.existsSync(`${testSrc}/repos/${repoList[id]}/`)) {
+        fs.mkdirSync(`${testSrc}/repos/${repoList[id]}/`, err => {
+          if (err) {
+            console.error(err)
+          }
+          // file written successfully
+        })
+      }
+      // Change directory path run git command
+      execSync('git --bare init', {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/repos/${repoList[id]}`, '') // where we're cloning the repo to
+      })
+      // Change directory path run git command
+      execSync(`git clone ${testSrc}/repos/${repoList[id]}`, {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/clones`, '') // where we're cloning the repo to
+      })
+      if (repoList[id] === 'repo16') {
+        fs.writeFileSync(`${testSrc}/clones/repo16/reposFile.json`, JSON.stringify(repo16FileData))
+      }
+      // Change directory path
+      process.chdir(`${testSrc}/clones/${repoList[id]}`)
+      // Create the package.json and package-lock.json file in the ./clones/..
+      fs.writeFileSync(`${testSrc}/clones/${repoList[id]}/package.json`, JSON.stringify(packageList[id][0]))
+      fs.writeFileSync(`${testSrc}/clones/${repoList[id]}/package-lock.json`, JSON.stringify(packageList[id][1]))
+
+      // Run git command to push package and package-lock files
+      execSync('git add .', {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/clones/${repoList[id]}`, '') // where we're cloning the repo to
+      })
+      execSync('git commit -m "commit"', {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/clones/${repoList[id]}`, '') // where we're cloning the repo to
+      })
+      execSync('git push', {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/clones/${repoList[id]}`, '') // where we're cloning the repo to
+      })
+    }
+    execSync('npm ci', {
+      stdio: 'pipe', // hide output from git
+      cwd: path.resolve(`${testSrc}/clones/repo16`, '') // where we're cloning the repo to
+    })
+    execSync('npm ci', {
+      stdio: 'pipe', // hide output from git
+      cwd: path.resolve(`${testSrc}/clones/repo16`, '') // where we're cloning the repo to
+    })
+  } catch {}
+}
