+++ test/util/fallbackSB09.js
@@ -1,0 +1,197 @@
+module.exports = fallbackDependancySandBox
+
+function fallbackDependancySandBox (appDir) {
+  const fs = require('fs')
+  const path = require('path')
+  const testSrc = path.resolve(__dirname, '../../test')
+  const { execSync } = require('child_process')
+  const repoList = ['repo25', 'repo26', 'repo27']
+
+  if (!appDir) {
+    let processEnv
+    if (fs.existsSync(path.join(process.cwd(), 'node_modules')) === false) {
+      processEnv = process.cwd()
+    } else {
+      processEnv = undefined
+    }
+    appDir = processEnv
+  }
+
+  // Checks if the repos folder exist, if not it creates it
+  try {
+    if (!fs.existsSync(`${testSrc}/repos`)) {
+      // creat the repos folder in the test folder
+      fs.mkdirSync(`${testSrc}/repos`, err => {
+        if (err) {
+          console.error(err)
+        }
+        // file written successfully
+      })
+    }
+    // Checks if the clone folder exist, if not it creates it
+    if (!fs.existsSync(`${testSrc}/clones`)) {
+      // creat the clones folder in the test folder
+      fs.mkdirSync(`${testSrc}/clones`, err => {
+        if (err) {
+          console.error(err)
+        }
+        // file written successfully
+      })
+    }
+
+    const repo25Package = {
+      devDependencies: {
+        'fallback-dependencies': '../../../'
+      },
+      fallbackDependencies: {
+        dir: 'lib',
+        reposFile: 'reposFile.json'
+      },
+      scripts: {
+        postinstall: 'node node_modules/fallback-dependencies/fallback-dependencies.js'
+      }
+    }
+    const repo25PackageLocked = {
+      name: 'repo25',
+      lockfileVersion: 3,
+      requires: true,
+      packages: {
+        '': {
+          hasInstallScript: true,
+          devDependencies: {
+            'fallback-dependencies': '../../../'
+          }
+        },
+        '../../..': {
+          version: '0.1.0',
+          dev: true,
+          license: 'CC-BY-25.0',
+          devDependencies: {}
+        },
+        'node_modules/fallback-dependencies': {
+          resolved: '../../..',
+          link: true
+        }
+      }
+    }
+    const repo25FileData = {
+      'fallback-deps-test-repo-26': [
+        '../../../repos/repo26'
+      ]
+    }
+    const repo26Package = {
+      devDependencies: {
+        'fallback-dependencies': '../../../../../../'
+      },
+      fallbackDependencies: {
+        dir: 'lib',
+        repos: {
+          'fallback-deps-test-repo-27': '../../../../../repos/repo27'
+        }
+      },
+      scripts: {
+        postinstall: 'node node_modules/fallback-dependencies/fallback-dependencies.js'
+      }
+    }
+    const repo26PackageLocked = {
+      name: 'repo25',
+      lockfileVersion: 3,
+      requires: true,
+      packages: {
+        '': {
+          hasInstallScript: true,
+          devDependencies: {
+            'fallback-dependencies': '../../../../../'
+          }
+        },
+        '../../../../..': {
+          version: '0.1.0',
+          dev: true,
+          license: 'CC-BY-25.0',
+          devDependencies: {}
+        },
+        'node_modules/fallback-dependencies': {
+          resolved: '../../../../..',
+          link: true
+        }
+      }
+    }
+    const repo27Package = {}
+    const repo27PackageLocked = {
+      name: 'repo27',
+      lockfileVersion: 3,
+      requires: true,
+      packages: {}
+    }
+    const packageList = [[repo25Package, repo25PackageLocked], [repo26Package, repo26PackageLocked], [repo27Package, repo27PackageLocked]]
+    for (const id in repoList) {
+      if (!fs.existsSync(`${testSrc}/repos/${repoList[id]}/`)) {
+        fs.mkdirSync(`${testSrc}/repos/${repoList[id]}/`, err => {
+          if (err) {
+            console.error(err)
+          }
+          // file written successfully
+        })
+      }
+      // Change directory path run git command
+      execSync('git --bare init', {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/repos/${repoList[id]}`, '') // where we're cloning the repo to
+      })
+      // Change directory path run git command
+      execSync(`git clone ${testSrc}/repos/${repoList[id]}`, {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/clones`, '') // where we're cloning the repo to
+      })
+      if (repoList[id] === 'repo25') {
+        fs.writeFileSync(`${testSrc}/clones/repo25/reposFile.json`, JSON.stringify(repo25FileData))
+      }
+      // Change directory path
+      process.chdir(`${testSrc}/clones/${repoList[id]}`)
+      // Create the package.json and package-lock.json file in the ./clones/..
+      fs.writeFileSync(`${testSrc}/clones/${repoList[id]}/package.json`, JSON.stringify(packageList[id][0]))
+      fs.writeFileSync(`${testSrc}/clones/${repoList[id]}/package-lock.json`, JSON.stringify(packageList[id][1]))
+      // Run git command to push package and package-lock files
+      execSync('git add .', {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/clones/${repoList[id]}`, '') // where we're cloning the repo to
+      })
+      execSync('git commit -m "commit"', {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/clones/${repoList[id]}`, '') // where we're cloning the repo to
+      })
+      execSync('git push', {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/clones/${repoList[id]}`, '') // where we're cloning the repo to
+      })
+    }
+
+    execSync('npm ci', {
+      stdio: 'pipe', // hide output from git
+      cwd: path.resolve(`${testSrc}/clones/repo25`, '') // where we're cloning the repo to
+    })
+
+    fs.rmSync(path.resolve(`${testSrc}/repos/repo26/config`), { recursive: true, force: true })
+    fs.rmSync(path.resolve(`${testSrc}/clones/repo26/.git`), { recursive: true, force: true })
+
+    fs.rmSync(path.resolve(`${testSrc}/clones/repo25/lib/fallback-deps-test-repo-26/.git`), { recursive: true, force: true })
+    try {
+      execSync('git add .', {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/clones/repo25`, '') // where we're cloning the repo to
+      })
+      execSync('git commit -m "commit"', {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/clones/repo25`, '') // where we're cloning the repo to
+      })
+      execSync('git push', {
+        stdio: 'pipe', // hide output from git
+        cwd: path.resolve(`${testSrc}/clones/repo25`, '') // where we're cloning the repo to
+      })
+    } catch {}
+    execSync('npm ci', {
+      stdio: 'pipe', // hide output from git
+      cwd: path.resolve(`${testSrc}/clones/repo25`, '') // where we're cloning the repo to
+    })
+  } catch {}
+}
